+++ ui/pages/settings/settings-search/settings-search.js
